hammerhead: QCamera2: Add HAL3 support for flashlight
Implemented new API at HAL, allowing framework to turn on flash
devices independently of the sensor. The primary logic is
implemented as a singleton named QCameraFlash, which tracks
the state of any flash devices and interfaces with the driver
to turn them on or off.

Change-Id: I4452e65401140e28aa86184800857a423c9c4b09
Adapted to hammerhead camera from CAF via shamu's
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>